+++ webapp/src/app/components/widgets/event-widget/event-registry/event-registry.service.ts
@@ -20,6 +20,7 @@ import { ShowChoice } from './show-choice';
 import { ShowMsg } from './show-msg';
 import { StartNpcTradeMenu } from './start-npc-trade-menu';
 import { Wait } from './wait';
+import { ShowSideMsg } from './show-side-msg';
 
 type EventConstructor<T extends EventType> = new (domSanitizer: DomSanitizer, data: T, actionStep: boolean) => AbstractEvent<T>;
 
@@ -50,6 +51,7 @@ export class EventRegistryService {
 		this.register('GOTO_LABEL', GotoLabel);
 		this.register('START_NPC_TRADE_MENU', StartNpcTradeMenu);
 		this.register('OPEN_QUEST_DIALOG', OpenQuestDialog);
+		this.register('SHOW_SIDE_MSG', ShowSideMsg);
 		
 	}