+++ webapp/src/app/components/widgets/event-widget/event-registry/show-side-msg.ts
@@ -1,0 +1,26 @@
+import { DefaultEvent } from './default-event';
+import { EventType } from './abstract-event';
+import { Label, Person } from '../../../../models/events';
+
+interface ShowSideMsgData extends EventType {
+	message: Label;
+	person: Person;
+}
+
+export class ShowSideMsg extends DefaultEvent<ShowSideMsgData> {
+	
+	override update() {
+		this.info = this.combineStrings(
+			this.getTypeString('#fa5aff'),
+			this.getColoredString(this.data.person?.person + '>&#8203;' + this.data.person?.expression, '#ffe188'),
+			this.data.message?.en_US
+		);
+	}
+	
+	protected override generateNewDataInternal() {
+		return {
+			person: {},
+			message: {}
+		};
+	}
+}