+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
+- Made Event `SHOW_SIDE_MSG` prettier. [image](https://github.com/CCDirectLink/crosscode-map-editor/assets/9483499/7aa7dce2-f678-4b93-a026-fbbc5873ebd0)
 - Double clicking an event now opens the event editor
 - Double clicking an npc now opens the npc state editor